+++ src/services/AssetVolumes.php
@@ -43,28 +43,16 @@ class AssetVolumes extends BaseMigration
         ];
 
         if ($volume->hasUrls){
-            $newVolume['hasUrls'] = 1;
-            $newVolume['url'] = $volume->url;
+          $newVolume['hasUrls'] = 1;
+          $newVolume['url'] = $volume->url;
         }
 
         if ($fullExport) {
-            $newVolume['fieldLayout'] = array();
-            $fieldLayout = $volume->getFieldLayout();
-
-            foreach ($fieldLayout->getTabs() as $tab) {
-                $newVolume['fieldLayout'][$tab->name] = array();
-                foreach ($tab->getFields() as $tabField) {
-
-                    $newVolume['fieldLayout'][$tab->name][] = $tabField->handle;
-                    if ($tabField->required) {
-                        $newVolume['requiredFields'][] =$tabField->handle;
-                    }
-                }
-            }
-         }
+          $this->getFieldLayout($volume->getFieldLayout(), $newVolume);
+        }
 
         if ($fullExport) {
-            $newVolume = $this->onBeforeExport($volume, $newVolume);
+          $newVolume = $this->onBeforeExport($volume, $newVolume);
         }
 
         return $newVolume;
@@ -116,7 +104,7 @@ class AssetVolumes extends BaseMigration
     public function createModel(Array $data)
     {
         $volumes = Craft::$app->getVolumes();
-        
+
         $volume = $volumes->createVolume([
             'id' => array_key_exists('id', $data) ? $data['id'] : null,
             'type' => $data['type'],
@@ -128,35 +116,12 @@ class AssetVolumes extends BaseMigration
             'settings' => $data['typesettings']
         ]);
 
-        if (array_key_exists('fieldLayout', $data)) {
-            $requiredFields = array();
-            if (array_key_exists('requiredFields', $data)) {
-                foreach ($data['requiredFields'] as $handle) {
-                    $field = Craft::$app->fields->getFieldByHandle($handle);
-                    if ($field) {
-                        $requiredFields[] = $field->id;
-                    }
-                }
-            }
-
-            $layout = array();
-            foreach ($data['fieldLayout'] as $key => $fields) {
-                $fieldIds = array();
-                foreach ($fields as $field) {
-                    $existingField = Craft::$app->fields->getFieldByHandle($field);
-                    if ($existingField) {
-                        $fieldIds[] = $existingField->id;
-                    } else {
-                        $this->addError('error', 'Missing field: '.$field.' can not add to field layout for Asset Volume: '.$volume->handle);
-                    }
-                }
-                $layout[$key] = $fieldIds;
-            }
-
-            $fieldLayout = Craft::$app->fields->assembleLayout($layout, $requiredFields);
-            $fieldLayout->type = Asset::class;
-            $volume->fieldLayout = $fieldLayout;
+        $fieldLayout = $this->createFieldLayout($data);
+        if ($fieldLayout) {
+          $fieldLayout->type = Asset::class;
+          $volume->setFieldLayout($fieldLayout);
         }
+
         return $volume;
     }