+++ src/services/Tags.php
@@ -38,18 +38,7 @@ class Tags extends BaseMigration
         $this->addManifest($tag->handle);
 
         if ($fullExport) {
-            $newTag['fieldLayout'] = array();
-            $newTag['requiredFields'] = array();
-            $fieldLayout = $tag->getFieldLayout();
-            foreach ($fieldLayout->getTabs() as $tab) {
-                $newTag['fieldLayout'][$tab->name] = array();
-                foreach ($tab->getFields() as $tabField) {
-                    $newTag['fieldLayout'][$tab->name][] = $tabField->handle;
-                    if ($tabField->required) {
-                        $newTag['requiredFields'][] = $tabField->handle;
-                    }
-                }
-            }
+          $this->getFieldLayout($tag->getFieldLayout(), $newTag);
         }
 
         if ($fullExport) {
@@ -114,35 +103,10 @@ class Tags extends BaseMigration
         $tag->name = $data['name'];
         $tag->handle = $data['handle'];
 
-        if (array_key_exists('fieldLayout', $data)) {
-            $requiredFields = array();
-            if (array_key_exists('requiredFields', $data)) {
-                foreach ($data['requiredFields'] as $handle) {
-                    $field = Craft::$app->fields->getFieldByHandle($handle);
-                    if ($field) {
-                        $requiredFields[] = $field->id;
-                    }
-                }
-            }
-
-            $layout = array();
-            foreach ($data['fieldLayout'] as $key => $fields) {
-                $fieldIds = array();
-                foreach ($fields as $field) {
-                    $existingField = Craft::$app->fields->getFieldByHandle($field);
-                    if ($existingField) {
-                        $fieldIds[] = $existingField->id;
-                    } else {
-                        $this->addError('error', 'Missing field: ' . $field . ' can not add to field layout for Tag: ' . $tag->handle);
-                    }
-                }
-                $layout[$key] = $fieldIds;
-            }
-
-
-            $fieldLayout = Craft::$app->fields->assembleLayout($layout, $requiredFields);
-            $fieldLayout->type = Tag::class;
-            $tag->fieldLayout = $fieldLayout;
+        $fieldLayout = $this->createFieldLayout($data);
+        if ($fieldLayout) {
+          $fieldLayout->type = Tag::class;
+          $tag->setFieldLayout($fieldLayout);
         }
 
         return $tag;